refactor: separate interlinks index responsibilities
